docs: Do not list pptx/ppt files. Converted to pdf when downloaded.
Since getting ppt metadata might become a challenge, we convert
these to pdf instead upon download. If a documents need to be
modified, then we can still open the original which is also saved
in the same directory. The conversion is currently done with Office's
"Save as" function. A command line to do batch conversion would be
nice.