File type overhaul: Use exiftool instead of PDFInfo.
The PDFinfo was fast but couldn't cope with different types of
metadata. It also always managed to find the wrong metadata while
other tools do a much better job.

exiftool supports many file types so it's better to use. However,
the current issue is it's way too slow. So that needs to be
addressed. Until it is addressed, we disable some of the documents
in the cpu_docs/ folder.

A side-fix that shouldn't be in this commit: top_dir was returning
junk based on really bad coding. Now it's better.