docs: Add details, fix top_dir again.
Details: Add a new column. When clicked this runs details.php
which is basically a shell for exiftool for that file. The file
is passed in file form field. The encoding/decoding is a b**ch.

urldecode converts spaces to + signs and + signs to %2B. urldecode
does not decode these though. We have to use rawurldecode which
decodes them properly somehow.

As for the top_dir, it's hurting my head but here's the situation.
Let's say current dir is: C:/A/B/C and the file is C:/A/B/C/D/E/file.pdf.
The top_dir should return "D". So here's what we do:

Look for "current dir" in the file's full path. Point the string
to the end of that. Then look for the first '/' character. That's
the end of our string. So get the substring from zero to what
we just looked for.