exif batch processing: Implemented.
We go over the file list and prune it (excludes). Then we create
a string of all files we want to process in a batch, separated by
double quotes and spaces. This is fed to exif which returns a nice
array of arrays in php format. For each file, we get the description
and the date (and some other data) and print our table.