Update.php: Add preliminary support to update the links.
This commit creates a database, db/docs.db, which contains
the following information:

	link, description, date, checksum

When the update function is called, we check if the sha1 of the
file exists in the current database. If so, there is no need to
update. If not, we add a new row with the relevant information.

So far, only the path name (no link) and the sha1 information
is valid. The rest will come later.