Update: Finish update procedure.
In this commit, we retrieve the description and the date from the
PDF file using exiftool and commit it to the database.

We also increased the database columns to include some more data.
The new format is:

	link, path, dir, filename, description, date, checksum

Single quotes are cleaned up from the description since they
give the sqlite exec statement trouble.