Update: Escape single quotes properly.
If the path or the link has a single quote, we cannot just remove
it since those will be used to find the file. So we escape it.
However, sqlite does not use the backslash escape method. It uses
a single quote to escape a single quote. So do that.