Update: Add trimming functionality.
During the update, we also trim the database and remove the entries
that correspond to files that are not found.

For this we add a new column called valid. This is set to 1 during
update if the sha1 has been matched to a file, or a new file has
been added to the DB. After update is done, we delete the entries
where valid is 0. We then set all valid entries to 0 for the next
time.