css: Remove underline from links.
When I adjusted the html code to align the links, it caused
the underline to extend beyond the word. It looked like this:

	linkname
	^^^^^^^^^^^           <- underlining whitespace

So to avoid this issue I'd either have to un-align the code or
just remove the underline from links. Let's remove it to make
it look snazzier anyway.

Change the link colors too while we're at it. All states of
the link get the same color, because I pretty much visit all these
links so there is no need to have different colors really.